Renamed unit_cell.py as crystal_lattice.py
